requirements.txt
streamlit>=1.25.0
google-generativeai>=0.5.2
python-dotenv>=1.0.1
reportlab
plotly